LK-Webhook/pyproject.toml
[project]
name = "lk-webhook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "livekit-agents>=1.3.10",
    "python-dotenv>=1.2.1",
]
